requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
fastapi==0.68.1
gunicorn==20.1.0
h11==0.12.0
numpy==1.21.2
opencv-python-headless==4.2.0.32
Pillow==8.3.2
pydantic==1.8.2
pytesseract==0.3.8
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
